# Sped up AutoEnum performance