--- a/.github/workflows/tests.yml
+++ b/.github/workflows/tests.yml
@@ -29,4 +29,4 @@ jobs:
 
       - name: Run tests
         run: |
-          pytest tests --verbose
+          pytest tests --verbose -s