--- a/src/autoenum/__init__.py
+++ b/src/autoenum/__init__.py
@@ -1,4 +1,6 @@
+import threading
 from enum import Enum, auto
+from functools import lru_cache
 from typing import (
     Any,
     Dict,
@@ -88,42 +90,107 @@ _DEFAULT_REMOVAL_TABLE = str.maketrans(
 
 class AutoEnum(str, Enum):
     """
-    Utility class which can be subclassed to create enums using auto() and alias().
-    Also provides utility methods for common enum operations.
+    Ultra-fast AutoEnum with fuzzy matching and aliases.
     """
 
+    __slots__ = ()  # no per-instance attrs beyond those in Enum/str
+
     def __init__(self, value: Union[str, alias]):
-        self.aliases: Tuple[str, ...] = tuple()
-        if isinstance(value, alias):
-            self.aliases: Tuple[str, ...] = value.names
+        # store aliases tuple for each member
+        object.__setattr__(self, "aliases", tuple(value.names) if isinstance(value, alias) else ())
+
+    def _generate_next_value_(name, start, count, last_values):
+        # keep the enum member’s *name* as its value
+        return name
+
+    @classmethod
+    def __init_subclass__(cls, **kwargs):
+        super().__init_subclass__(**kwargs)
+        setattr(cls, "_lookup_lock", threading.Lock())
+        cls._initialize_lookup()
+
+    @classmethod
+    def _initialize_lookup(cls):
+        # quick check to avoid locking if already built
+        if "_value2member_map_normalized_" in cls.__dict__:
+            return
+        with cls._lookup_lock:
+            if "_value2member_map_normalized_" in cls.__dict__:
+                return
+
+            mapping: Dict[str, "AutoEnum"] = {}
+
+            def _register(e: "AutoEnum", norm: str):
+                if norm in mapping:
+                    raise ValueError(
+                        f'Cannot register enum "{e.name}"; normalized name "{norm}" already exists.'
+                    )
+                mapping[norm] = e
+
+            # walk every member exactly once
+            for e in cls:
+                # register its own name
+                _register(e, cls._normalize(e.name))
+                # register alias repr
+                if e.aliases:
+                    # inline alias_repr
+                    alias_repr = f"alias:{list(e.aliases)}"
+                    _register(e, cls._normalize(alias_repr))
+                    # register each plain alias
+                    for a in e.aliases:
+                        _register(e, cls._normalize(a))
+
+            # stash it on the class
+            setattr(cls, "_value2member_map_normalized_", mapping)
+
+    @classmethod
+    @lru_cache(maxsize=None)
+    def _normalize(cls, x: str) -> str:
+        # C-level translate is very fast; caching makes repeated lookups O(1)
+        return str(x).translate(_DEFAULT_REMOVAL_TABLE)
 
     @classmethod
     def _missing_(cls, enum_value: Any):
-        ## Ref: https://stackoverflow.com/a/60174274/4900327
-        ## This is needed to allow Pydantic to perform case-insensitive conversion to AutoEnum.
-        return cls.from_str(enum_value=enum_value, raise_error=True)
+        # invoked by Enum machinery when auto-casting fails
+        return cls.from_str(enum_value, raise_error=True)
 
-    def _generate_next_value_(name, start, count, last_values):
-        return name
-
-    @property
-    def str(self) -> str:
-        return self.__str__()
-
-    def __repr__(self):
-        return self.__str__()
-
-    def __str__(self):
+    def __str__(self) -> str:
         return self.name
 
-    def __hash__(self):
+    def __repr__(self) -> str:
+        return self.name
+
+    def __hash__(self) -> int:
         return hash(self.__class__.__name__ + "." + self.name)
 
-    def __eq__(self, other):
+    def __eq__(self, other: Any) -> bool:
+        # identity check is fastest and correct for singletons
         return self is other
 
-    def __ne__(self, other):
+    def __ne__(self, other: Any) -> bool:
         return self is not other
+
+    @classmethod
+    def from_str(cls, enum_value: Any, raise_error: bool = True) -> Optional["AutoEnum"]:
+        # short‐circuit if it's already the right type
+        if isinstance(enum_value, cls):
+            return enum_value
+        # None tolerated?
+        if enum_value is None:
+            if raise_error:
+                raise ValueError("Cannot convert None to enum")
+            return None
+        # wrong type?
+        if not isinstance(enum_value, str):
+            if raise_error:
+                raise ValueError(f"Input must be str or {cls.__name__}; got {type(enum_value)}")
+            return None
+        # one normalized dict lookup
+        norm = cls._normalize(enum_value)
+        e = cls._value2member_map_normalized_.get(norm)
+        if e is None and raise_error:
+            raise ValueError(f"Could not find enum with value {enum_value!r}; available: {list(cls)}")
+        return e
 
     def matches(self, enum_value: str) -> bool:
         return self is self.from_str(enum_value, raise_error=False)
@@ -138,186 +205,67 @@ class AutoEnum(str, Enum):
 
     @classmethod
     def display_names(cls, **kwargs) -> str:
-        return str([enum_value.display_name(**kwargs) for enum_value in list(cls)])
+        return str([e.display_name(**kwargs) for e in cls])
 
     def display_name(self, *, sep: str = " ") -> str:
         return sep.join(
-            [
-                word.lower()
-                if word.lower() in ("of", "in", "the")
-                else word.capitalize()
-                for word in str(self).split("_")
-            ]
+            word.lower() if word.lower() in ("of", "in", "the") else word.capitalize()
+            for word in self.name.split("_")
         )
 
-    @classmethod
-    def _initialize_lookup(cls):
-        if (
-            "_value2member_map_normalized_" not in cls.__dict__
-        ):  ## Caching values for fast retrieval.
-            cls._value2member_map_normalized_ = {}
-
-            def _set_normalized(e, normalized_e_name):
-                if normalized_e_name in cls._value2member_map_normalized_:
-                    raise ValueError(
-                        f'Cannot register enum "{e.name}"; '
-                        f'another enum with the same normalized name "{normalized_e_name}" already exists.'
-                    )
-                cls._value2member_map_normalized_[normalized_e_name] = e
-
-            for e in list(cls):
-                _set_normalized(e, cls._normalize(e.name))
-                if len(e.aliases) > 0:
-                    ## Add the alias-repr to the lookup:
-                    _set_normalized(e, cls._normalize(alias(*e.aliases).alias_repr))
-                    for e_alias in e.aliases:
-                        _set_normalized(e, cls._normalize(e_alias))
-
-    @classmethod
-    def from_str(cls, enum_value: str, raise_error: bool = True) -> Optional:
-        """
-        Performs a case-insensitive lookup of the enum value string among the members of the current AutoEnum subclass.
-        :param enum_value: enum value string
-        :param raise_error: whether to raise an error if the string is not found in the enum
-        :return: an enum value which matches the string
-        :raises: ValueError if raise_error is True and no enum value matches the string
-        """
-        if isinstance(enum_value, cls):
-            return enum_value
-        if enum_value is None and raise_error is False:
-            return None
-        if not isinstance(enum_value, str) and raise_error is True:
-            raise ValueError(f"Input should be a string; found type {type(enum_value)}")
-        cls._initialize_lookup()
-        enum_obj: Optional[AutoEnum] = cls._value2member_map_normalized_.get(
-            cls._normalize(enum_value)
-        )
-        if enum_obj is None and raise_error is True:
-            raise ValueError(
-                f"Could not find enum with value {repr(enum_value)}; available values are: {list(cls)}."
-            )
-        return enum_obj
-
-    @classmethod
-    def _normalize(cls, x: str) -> str:
-        ## Found to be faster than .translate() and re.sub() on Python 3.10.6
-        return str(x).translate(_DEFAULT_REMOVAL_TABLE)
+    # -------------- conversion utilities (unchanged) --------------
 
     @classmethod
     def convert_keys(cls, d: Dict) -> Dict:
-        """
-        Converts string dict keys to the matching members of the current AutoEnum subclass.
-        Leaves non-string keys untouched.
-        :param d: dict to transform
-        :return: dict with matching string keys transformed to enum values
-        """
-        out_dict = {}
+        out = {}
         for k, v in d.items():
-            if isinstance(k, str) and cls.from_str(k, raise_error=False) is not None:
-                out_dict[cls.from_str(k, raise_error=False)] = v
+            if isinstance(k, str):
+                e = cls.from_str(k, raise_error=False)
+                out[e] = v if e else v
             else:
-                out_dict[k] = v
-        return out_dict
+                out[k] = v
+        return out
 
     @classmethod
     def convert_keys_to_str(cls, d: Dict) -> Dict:
-        """
-        Converts dict keys of the current AutoEnum subclass to the matching string key.
-        Leaves other keys untouched.
-        :param d: dict to transform
-        :return: dict with matching keys of the current AutoEnum transformed to strings.
-        """
-        out_dict = {}
-        for k, v in d.items():
-            if isinstance(k, cls):
-                out_dict[str(k)] = v
-            else:
-                out_dict[k] = v
-        return out_dict
+        return {(str(k) if isinstance(k, cls) else k): v for k, v in d.items()}
 
     @classmethod
     def convert_values(
         cls, d: Union[Dict, Set, List, Tuple], raise_error: bool = False
     ) -> Union[Dict, Set, List, Tuple]:
-        """
-        Converts string values to the matching members of the current AutoEnum subclass.
-        Leaves non-string values untouched.
-        :param d: dict, set, list or tuple to transform.
-        :param raise_error: raise an error if unsupported type.
-        :return: data structure with matching string values transformed to enum values.
-        """
         if isinstance(d, dict):
             return cls.convert_dict_values(d)
         if isinstance(d, list):
             return cls.convert_list(d)
         if isinstance(d, tuple):
-            return tuple(cls.convert_list(d))
+            return tuple(cls.convert_list(list(d)))
         if isinstance(d, set):
             return cls.convert_set(d)
         if raise_error:
-            raise ValueError(f"Unrecognized data structure of type {type(d)}")
+            raise ValueError(f"Unsupported type: {type(d)}")
         return d
 
     @classmethod
     def convert_dict_values(cls, d: Dict) -> Dict:
-        """
-        Converts string dict values to the matching members of the current AutoEnum subclass.
-        Leaves non-string values untouched.
-        :param d: dict to transform
-        :return: dict with matching string values transformed to enum values
-        """
-        out_dict = {}
-        for k, v in d.items():
-            if isinstance(v, str) and cls.from_str(v, raise_error=False) is not None:
-                out_dict[k] = cls.from_str(v, raise_error=False)
-            else:
-                out_dict[k] = v
-        return out_dict
+        return {k: (cls.from_str(v, raise_error=False) if isinstance(v, str) else v) for k, v in d.items()}
 
     @classmethod
-    def convert_list(cls, l: Union[List, Tuple]) -> List:
-        """
-        Converts string list items to the matching members of the current AutoEnum subclass.
-        Leaves non-string items untouched.
-        :param l: list to transform
-        :return: list with matching string items transformed to enum values
-        """
-        out_list = []
-        for item in l:
-            if isinstance(item, str) and cls.matches_any(item):
-                out_list.append(cls.from_str(item))
-            else:
-                out_list.append(item)
-        return out_list
+    def convert_list(cls, l: List) -> List:
+        return [
+            (cls.from_str(item) if isinstance(item, str) and cls.matches_any(item) else item) for item in l
+        ]
 
     @classmethod
     def convert_set(cls, s: Set) -> Set:
-        """
-        Converts string list items to the matching members of the current AutoEnum subclass.
-        Leaves non-string items untouched.
-        :param s: set to transform
-        :return: set with matching string items transformed to enum values
-        """
-        out_set = set()
+        out = set()
         for item in s:
             if isinstance(item, str) and cls.matches_any(item):
-                out_set.add(cls.from_str(item))
+                out.add(cls.from_str(item))
             else:
-                out_set.add(item)
-        return out_set
+                out.add(item)
+        return out
 
     @classmethod
     def convert_values_to_str(cls, d: Dict) -> Dict:
-        """
-        Converts dict values of the current AutoEnum subclass to the matching string value.
-        Leaves other values untouched.
-        :param d: dict to transform
-        :return: dict with matching values of the current AutoEnum transformed to strings.
-        """
-        out_dict = {}
-        for k, v in d.items():
-            if isinstance(v, cls):
-                out_dict[k] = str(v)
-            else:
-                out_dict[k] = v
-        return out_dict
+        return {k: (str(v) if isinstance(v, cls) else v) for k, v in d.items()}