## Before speedup (Python 3.11.12, M4 Macbook Air):
[from_str_lookup]               1M iterations took 420.18ms (avg 0.42us per from_str lookup)
[no_cache_from_str_lookup]      10k iterations took 59.82ms (avg 5.98us per from_str lookup without cache)
[matches_any_lookup]            1M iterations took 492.67ms (avg 0.49us per matches_any lookup)
[no_cache_matches_any_lookup]   10k iterations took 54.57ms (avg 5.46us per matches_any lookup without cache)
[enum_iteration]                1M enum-iterations took 3481.55ms  (avg 3.48us per enum-iterations)
[convert_list]                  1M list-conversions took 3197.16ms (avg 3.20us per list-conversion)

## After speedup (Python 3.11.12, M4 Macbook Air):
[from_str_lookup]               1M iterations took 219.85ms (avg 0.22us per from_str lookup)
[no_cache_from_str_lookup]      10k iterations took 4.15ms  (avg 0.41us per from_str lookup without cache)
[matches_any_lookup]            1M iterations took 303.36ms (avg 0.30us per matches_any lookup)
[no_cache_matches_any_lookup]   10k iterations took 7.84ms  (avg 0.78us per matches_any lookup without cache)
[enum_iteration]                1M enum-iterations took 3452.85ms  (avg 3.45us per enum-iterations)
[convert_list]                  1M list-conversions took 2004.54ms (avg 2.00us per list-conversion)

===============================================================================

## Before speedup (Python 3.12.11, M4 Macbook Air):
[from_str_lookup]               1M iterations took 413.97ms (avg 0.41us per from_str lookup)
[no_cache_from_str_lookup]      10k iterations took 58.06ms (avg 5.81us per from_str lookup without cache)
[matches_any_lookup]            1M iterations took 521.21ms (avg 0.52us per matches_any lookup)
[no_cache_matches_any_lookup]   10k iterations took 55.04ms (avg 5.50us per matches_any lookup without cache)
[enum_iteration]                1M enum-iterations took 3458.80ms  (avg 3.46us per enum-iterations)
[convert_list]                  1M list-conversions took 3217.21ms (avg 3.22us per list-conversion)

## After speedup (Python 3.12.11, M4 Macbook Air):
[from_str_lookup]               1M iterations took 158.65ms (avg 0.16us per from_str lookup)
[no_cache_from_str_lookup]      10k iterations took 3.55ms  (avg 0.36us per from_str lookup without cache)
[matches_any_lookup]            1M iterations took 212.31ms (avg 0.21us per matches_any lookup)
[no_cache_matches_any_lookup]   10k iterations took 7.58ms  (avg 0.76us per matches_any lookup without cache)
[enum_iteration]                1M enum-iterations took 2263.28ms  (avg 2.26us per enum-iterations)
[convert_list]                  1M list-conversions took 1283.08ms (avg 1.28us per list-conversion)